# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "THE METAL EDGE",
        "description": "The Metal Edge API",
        "termsOfService": "",
        "contact": {
            "email": "sumersingh1997.ssh@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add-product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add Product",
                "operationId": "productStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "discount_price": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/view-product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View product",
                "operationId": "productShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/edit-product/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Edit Product",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "discount_price": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "category": {
                                        "type": "text"
                                    },
                                    "subcategory": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/edit-product-status/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Edit Product Status",
                "operationId": "updateProductStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        },
        "/api/delete-product/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product",
                "operationId": "productDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "406": {
                        "description": "not acceptable",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_security": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_security": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer < token >)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}